Files for multi-select property moved to separate folder.
